+++ packages/boilerplate/composables/src/composables/getters/index.ts
@@ -6,6 +6,7 @@ import facetGetters from './facetGetters';
 import productGetters from './productGetters';
 import userGetters from './userGetters';
 import userShippingGetters from './userShippingGetters';
+import userBillingGetters from './userBillingGetters';
 import orderGetters from './orderGetters';
 import wishlistGetters from './wishlistGetters';
 
@@ -17,6 +18,7 @@ export {
   productGetters,
   userGetters,
   userShippingGetters,
+  userBillingGetters,
   orderGetters,
   wishlistGetters
 };