+++ packages/commercetools/api-client/src/api/getCart/index.ts
@@ -1,10 +1,10 @@
-import { apolloClient, getSettings } from './../../index';
-import { CartQueryResponse } from './../../types/Api';
+import { getSettings } from '../../index';
+import { CartQueryResponse } from '../../types/Api';
 import defaultQuery from './defaultQuery';
 
 const getCart = async (cartId: string): Promise<CartQueryResponse> => {
-  const { locale, acceptLanguage } = getSettings();
-  return await apolloClient.query({
+  const { locale, acceptLanguage, client } = getSettings();
+  return await client.query({
     query: defaultQuery,
     variables: { cartId,
       locale,