+++ packages/commercetools/composables/src/useCheckout/createSetPersonalDetails.ts
@@ -3,8 +3,9 @@
 import { updateCart, cartActions } from '@vue-storefront/commercetools-api';
 import { personalDetails, loading} from './shared';
 import initFields from './initFields';
+import { CustomQuery } from '@vue-storefront/core';
 
-const createSetPersonalDetails = ({ factoryParams, setShippingDetails, cartFields, setCart }) => async (data, options: any = {}) => {
+const createSetPersonalDetails = ({ factoryParams, setShippingDetails, cartFields, setCart }) => async (data, options: any = {}, customQuery?: CustomQuery) => {
   personalDetails.value = { ...personalDetails.value, ...data };
   const { firstName, lastName } = personalDetails.value;
 
@@ -17,7 +18,7 @@ const createSetPersonalDetails = ({ factoryParams, setShippingDetails, cartField
     actions: [
       cartActions.setCustomerEmail(personalDetails.value.email)
     ]
-  });
+  }, customQuery);
 
   setCart(cartResponse.data.cart);
   initFields(cartResponse.data.cart);