+++ packages/core/core/src/server/index.ts
@@ -1,4 +1,3 @@
-
 import express, { json } from 'express';
 
 const app = express();