+++ packages/core/nuxt-theme-module/theme/components/UserShippingAddress.vue
@@ -1,0 +1,9 @@
+<template>
+  <b>Please implement vendor specific UserShippingAddress component in 'components' directory</b>
+</template>
+
+<script>
+export default {
+  name: 'UserAddress'
+};
+</script>